Extract input value formatting components
