+++ src/components/openapi/OperationInputValue.module.css
@@ -1,0 +1,36 @@
+.parameterListHeader {
+  display: flex;
+}
+
+.parameterName {
+  font-weight: bold;
+  font-family: var(--ifm-font-family-monospace);
+  margin-right: 0.5rem;
+}
+
+.parameterType:before {
+  content: ":";
+  margin-right: 0.5rem;
+}
+
+.parameterType {
+  color: var(--ifm-color-gray-600);
+}
+
+.parameterListHeaderSpacer {
+  flex-grow: 1;
+  margin-top: 5px;
+  margin-left: 0.5rem;
+  margin-right: 0.5rem;
+  height: 0.5rem;
+  border-bottom: thin solid var(--ifm-color-gray-300);
+}
+
+.parameterListBody {
+  font-size: 0.8rem;
+}
+
+.parameterAlternative {
+  color: var(--ifm-color-gray-600);
+  font-weight: bold;
+}
