+++ src/components/openapi/OperationReference.module.css
@@ -25,46 +25,9 @@
   vertical-align: top;
 }
 
-.parameterListHeader {
-  display: flex;
-}
-
-.parameterListHeaderSpacer {
-  flex-grow: 1;
-  margin-top: 5px;
-  margin-left: 0.5rem;
-  margin-right: 0.5rem;
-  height: 0.5rem;
-  border-bottom: thin solid var(--ifm-color-gray-300);
-}
-
-.parameterName {
-  font-weight: bold;
-  font-family: var(--ifm-font-family-monospace);
-  margin-right: 0.5rem;
-}
-
-.parameterType:before {
-  content: ":";
-  margin-right: 0.5rem;
-}
-
-.parameterType {
-  color: var(--ifm-color-gray-600);
-}
-
-.parameterListBody {
-  font-size: 0.8rem;
-}
-
 .parameterExample {
   color: var(--ifm-color-gray-600);
   text-transform: uppercase;
   font-weight: bold;
   vertical-align: baseline;
-}
-
-.parameterAlternative {
-  color: var(--ifm-color-gray-600);
-  font-weight: bold;
 }
